blogTI/serializers.py
from rest_framework import serializers
from .models import Publicacion, Etiqueta, Reposteo, Comentario
from django.contrib.auth.models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'password', 'email']

class PublicacionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Publicacion
        fields = '__all__'

class EtiquetaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Etiqueta
        fields = '__all__'

class ComentarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comentario
        fields = '__all__'

class ReposteoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reposteo
        fields = '__all__'
